scripts/loaders/requirements.txt
psycopg2-binary==2.9.9
watchdog==4.0.0
